CHORE: Refactor sqs embedding task producer lambda to use config layer and filter utils to function that we might use in other lambdas
