+++ lambda/fetch_videos/config.json
@@ -1,3 +1,4 @@
 {
-  "presigned_url_expiry": 86400
+  "presigned_url_expiry": 86400,
+  "secret_name": "KubrickEmbeddingTaskConsumerSecret"
 }
