+++ lambda/sqs_embedding_task_producer/lambda_function.py
@@ -1,39 +1,84 @@
 import json
 import boto3
+import time
+import botocore.exceptions
 import os
 import logging
 import utils
+import urllib.parse
 from twelvelabs import TwelveLabs
-from config import load_config, get_secret
+from config import load_config, get_secret, setup_logging, get_db_config
 from vector_db_service import VectorDBService
 
-s3 = boto3.client("s3")
-sqs = boto3.client("sqs")
 
-logger = logging.getLogger()
-logger.setLevel(logging.INFO)
+def wait_for_file(
+    s3_client, bucket, key, retries=3, delay=2, logger=logging.getLogger()
+):
+    for attempt in range(retries):
+        try:
+            logger.info(
+                f"Checking if file exists: s3://{bucket}/{key} (attempt {attempt + 1}/{retries})"
+            )
+            s3_client.head_object(Bucket=bucket, Key=key)
+            logger.info("File found in S3")
+            return True
+        except botocore.exceptions.ClientError as e:
+            if e.response.get("Error", {}).get("Code") == "404":
+                logger.warning(f"File not found. Retrying in {delay} seconds...")
+                time.sleep(delay)
+            else:
+                logger.error(f"Unexpected error while checking file existence: {e}")
+                raise
+    logger.warning("File still not found after all retries")
+    return False
+
+
+def extract_s3_info(event):
+    records = event.get("Records", [])
+    if not records:
+        raise ValueError("No Records found in event")
+
+    record = records[0]
+    bucket = record.get("s3", {}).get("bucket", {}).get("name")
+    key = urllib.parse.unquote_plus(
+        record.get("s3", {}).get("object", {}).get("key", "")
+    ).strip()
+
+    if not bucket or not key:
+        raise ValueError("Missing bucket or key in event")
+
+    return bucket, key
+
+
+def persist_task_metadata(
+    db, metadata, fallback_status="failed", logger=logging.getLogger()
+):
+    try:
+        db.store_task(metadata)
+        logger.info(f"Task metadata stored in DB: {metadata}")
+    except Exception as e:
+        logger.error(f"Failed to store task metadata: {e}")
+        fallback = dict(metadata)
+        fallback["status"] = fallback_status
+        logger.error(f"Fallback data: {json.dumps(fallback)}")
 
 
 def lambda_handler(event, context):
-    logger.info("Lambda handler invoked")
-
+    s3 = boto3.client("s3")
+    sqs = boto3.client("sqs")
+    logger = setup_logging()
     config = load_config()
     SECRET = get_secret(config)
-
+    DB_CONFIG = get_db_config(SECRET)
     QUEUE_URL = os.environ["QUEUE_URL"]
-    DB_CONFIG = {
-        "host": os.getenv("DB_HOST", "localhost"),
-        "database": os.getenv("DB_NAME", "kubrick"),
-        "user": os.getenv("DB_USER", "postgres"),
-        "password": SECRET["DB_PASSWORD"],
-        "port": 5432,
-    }
+
+    logger.info("Lambda handler invoked")
 
     tl_client = TwelveLabs(api_key=SECRET["TWELVELABS_API_KEY"])
-    db = VectorDBService(db_params=DB_CONFIG, logger=logger)
+    vector_db_service = VectorDBService(db_params=DB_CONFIG, logger=logger)
 
     try:
-        bucket, key = utils.extract_s3_info(event)
+        bucket, key = extract_s3_info(event)
         logger.info(f"Extracted S3 bucket: {bucket}, key: {key}")
 
         if key.endswith("/"):
@@ -44,12 +89,13 @@ def lambda_handler(event, context):
             raise ValueError("File is not a video or the video format is not supported")
         logger.info("Validated video file type")
 
-        if not utils.wait_for_file(
+        if not wait_for_file(
             s3,
             bucket,
             key,
             config["file_check_retries"],
             config["file_check_delay_sec"],
+            logger,
         ):
             raise FileNotFoundError(f"File s3://{bucket}/{key} not found after retries")
 
@@ -84,7 +130,7 @@ def lambda_handler(event, context):
             "status": "processing",
             "twelvelabs_task_id": task_id,
         }
-        utils.persist_task_metadata(db, metadata)
+        persist_task_metadata(vector_db_service, metadata, logger=logger)
 
         logger.info("Lambda execution completed successfully")
         return {
@@ -104,6 +150,8 @@ def lambda_handler(event, context):
             "s3_key": locals().get("key"),
             "status": "failed",
         }
-        utils.persist_task_metadata(db, metadata, fallback_status="failed")
+        persist_task_metadata(
+            vector_db_service, metadata, fallback_status="failed", logger=logger
+        )
 
         return {"status": "error", "message": str(e)}
